breakup realtime.test.ts into 3 smaller tests
